requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech==1.32.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
Pillow==10.1.0